Move assets file code to more fitting locations
