+++ lib/config.js
@@ -3,7 +3,7 @@ const { klona } = require('klona');
 const { merge } = require('merge-anything');
 const { extname } = require('path');
 const { fileExtRegexStr, makeScopedClass, isDev } = require('./utils');
-const { asset, makeAssetsMap } = require('./assets');
+const { makeAssetsMap } = require('./esbuild');
 const postcss = require('esbuild-postcss');
 const browserslistToEsbuild = require('browserslist-to-esbuild');
 // const babel = require('esbuild-plugin-babel');
@@ -42,6 +42,18 @@ const buildShortcut = (shortcuts, { dir, alias = (name) => name }) => {
 
 
 module.exports = (opts = {}) => {
+  /*
+  custom beard asset tag render
+  */
+  function asset(opts) {
+    return (origPath, scopedOpts = {}) => {
+      const { content, path } = opts.map[origPath];
+      return scopedOpts.inline
+        ? content
+        : path;
+    }
+  }
+
   /*
   base component tag render
   - no clone of called component locals
