+++ lib/index.js
@@ -1,7 +1,7 @@
 const callsites = require('callsites');
 const beard = require('beard');
 const config = require('./config');
-const esbuild = require('./esbuild');
+const { esbuild } = require('./esbuild');
 const { dirname, resolve } = require('path');
 const { makeScopedClass } = require('./utils');
 const { bundle } = require('./bundling');
